// client/src/types.ts
export type BarMetadata = {
  _id: string
  yelpId: string
  address: string
  barName: string
  googleMapsLink: string
  happyHourDays: string
  happyHourHours: string
  happyHourMenu: string
  latitude: number
  longitude: number
  indoor?: string
  outdoor?: string
  monday?: string
  tuesday?: string
  wednesday?: string
  thursday?: string
  friday?: string
}